refactor: CloudFront에서 ARecord를 생성하는 것을 FrontendStack으로 변경
