+++ lib/cdn/SPACloudFront.ts
@@ -1,7 +1,7 @@
 import { Construct } from "constructs";
 import { Bucket } from "aws-cdk-lib/aws-s3";
 import { Certificate } from "aws-cdk-lib/aws-certificatemanager";
-import { ARecord, HostedZone, RecordTarget } from "aws-cdk-lib/aws-route53";
+import { IHostedZone } from "aws-cdk-lib/aws-route53";
 import {
   AccessLevel,
   Distribution,
@@ -9,17 +9,17 @@ import {
 } from "aws-cdk-lib/aws-cloudfront";
 import { S3BucketOrigin } from "aws-cdk-lib/aws-cloudfront-origins";
 import { Duration } from "aws-cdk-lib";
-import { CloudFrontTarget } from "aws-cdk-lib/aws-route53-targets";
 
 interface SPACloudFrontProps {
   environment: string;
   bucket: Bucket;
   certificateArn: string;
-  hostedZoneId: string;
-  hostedZoneName: string;
+  hostedZone: IHostedZone;
 }
 
 export class SPACloudFront extends Construct {
+  public readonly distribution: Distribution;
+
   constructor(scope: Construct, props: SPACloudFrontProps) {
     super(scope, `Web-CF-${props.environment}`);
 
@@ -29,16 +29,7 @@ export class SPACloudFront extends Construct {
       props.certificateArn,
     );
 
-    const hostedZone = HostedZone.fromHostedZoneAttributes(
-      this,
-      `HostedZone-${props.environment}`,
-      {
-        hostedZoneId: props.hostedZoneId,
-        zoneName: props.hostedZoneName,
-      },
-    );
-
-    const distribution = new Distribution(
+    this.distribution = new Distribution(
       this,
       `Distribution-${props.environment}`,
       {
@@ -64,17 +55,11 @@ export class SPACloudFront extends Construct {
           },
         ],
         domainNames: [
-          `www.${props.environment == "prod" ? "" : props.environment + "."}${hostedZone.zoneName}`,
+          `www.${props.environment == "prod" ? "" : props.environment + "."}${props.hostedZone.zoneName}`,
         ],
         certificate: certificate,
         comment: `Web Distribution ${props.environment}`,
       },
     );
-
-    new ARecord(this, `Web-ARecord-${props.environment}`, {
-      zone: hostedZone,
-      recordName: `www${props.environment == "prod" ? "" : `.${props.environment}`}`,
-      target: RecordTarget.fromAlias(new CloudFrontTarget(distribution)),
-    });
   }
 }
